+++ src/main/java/org/example/fibonacci/FibonacciStrategy.java
@@ -1,5 +1,7 @@
 package org.example.fibonacci;
 
+import java.math.BigInteger;
+
 public interface FibonacciStrategy {
-    long solveFibonacci(int n);
+    BigInteger solveFibonacci(int n);
 }
